Fixed repo again

+++ FYS4460/proj1/qtsrc/proj1/integrator.h
@@ -15,11 +15,11 @@ class Verlet_integrator : public Integrator {
     using Integrator :: Integrator;
 
 private:
-    void kick(float dt);
-    void drift(float dt);
+    void kick(double dt);
+    void drift(double dt);
 public:
-    void velocity_verlet_step(float dt);
-    void position_verlet_step(float dt);
+    void velocity_verlet_step(double dt);
+    void position_verlet_step(double dt);
 };
 
 #endif // INTEGRATOR_H
